inventory_api/docker/requirements.txt
hvac
flask
psycopg2-binary